Use standard metadata enrichment for Deezer uploads
Remove the custom sync refresh and completion markers. Existing uploads receive artwork through metadata enrichment; getter cache versioning prevents stale parsed items from hiding it.

Share grouped album metadata between listings and getters, preserve each requested album ID, and keep track and album image lists independent. Cover enrichment, cache invalidation and normal imports with regression tests.
